tests/wpt/web-platform-tests/trusted-types/Document-write.tentative.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support/helper.sub.js"></script>
<body>
<script>
  async_test(t => {
    createHTML_policy(window)
      .then(t.step_func_done(p => {
        let html = p.createHTML(INPUTS.HTML);
        document.write(html);
        assert_equals(document.body.innerText, RESULTS.HTML);
    }));
  }, "document.write with html assigned via policy (successful URL transformation).");
</script>
